formatting fixes

+++ src/lib.rs
@@ -596,7 +596,7 @@ fn execute_instr(mut input: InstrContext) -> Option<InstrEffects> {
             callbackback(&bincode::serialize(&Rent::default()).unwrap());
         }
         if *pubkey == last_restart_slot::id() {
-            let slot_val = 5000 as u64;
+            let slot_val = 5000_u64;
             callbackback(&bincode::serialize(&slot_val).unwrap());
         }
     });
